Moved JNIOffsets to x64.jni
